+++ src/rule/PseudoRule.ts
@@ -1,0 +1,13 @@
+import {createStyleRule, StyleRule} from "./StyleRule";
+
+const pseudo = createStyleRule<string>("pseudo", {
+  check(options, session) {
+    return Boolean(session.context.pseudoClasses?.includes(options));
+  },
+});
+
+export type PseudoRule = StyleRule & { type: string };
+
+export const createPseudoRule = (type: string): PseudoRule => Object.assign(() => pseudo(type), {type});
+
+export default pseudo;