+++ src/rule/index.ts
@@ -1,4 +1,5 @@
-export {default as media}    from "./MediaRule";
-export {default as platform} from "./PlatformRule";
+export {default as media}                    from "./MediaRule";
+export {default as platform}                 from "./PlatformRule";
+export {default as pseudo, createPseudoRule} from "./PseudoRule";
 
 export {and, or, createStyleRule} from "./StyleRule";